added "struct.h"
"struct.h" reusable header file. without conflicting with any other files this can be added as #include "struct.h"
