+++ hide/.gitignore
@@ -1,0 +1,13 @@
+
+fileManage.h
+filePrintFormat.h
+*.json
+*.h
+*.h
+printTheformat.c
+fileManage.h
+reff_mage-for_project_output.jpeg
+temp.c
+tempMain.c
+test.txt
+test2.txt
