+++ struct.h
@@ -1,0 +1,12 @@
+#ifndef STRUCT_H
+#define STRUCT_H
+
+
+typedef struct Info {
+    char name[30];
+    char ID[10];
+    int  units;
+    double amount;
+}Info;
+
+#endif
